+++ app/http/server.go
@@ -10,6 +10,7 @@ import (
 	"os/signal"
 	"time"
 
+	"github.com/labstack/echo-contrib/prometheus"
 	"github.com/labstack/echo/v4"
 )
 
@@ -28,13 +29,28 @@ func Start() {
 	dg.GET("/rapidoc", echo.WrapHandler(middlewares.RapiDocs()))
 	e.Static("/swagger.yaml", "./swagger.yaml")
 
-	container.Init(e.Group("api"))
+	// Create Prometheus server and Middleware
+	echoPrometheus := echo.New()
+	echoPrometheus.HideBanner = true
+	prom := prometheus.NewPrometheus("echo", nil)
 
-	port := config.App().Port
+	// Scrape metrics from Main Server
+	e.Use(prom.HandlerFunc)
+	// Setup metrics endpoint at another server
+	prom.SetMetricsPath(echoPrometheus)
+
+	go func() {
+		echoPrometheus.Logger.Fatal(echoPrometheus.Start(":" + config.App().MetricsPort))
+
+		// graceful shutdown prometheus server
+		GracefulShutdown(echoPrometheus)
+	}()
+
+	container.Init(e.Group("api"))
 
 	// start http server
 	go func() {
-		e.Logger.Fatal(e.Start(":" + port))
+		e.Logger.Fatal(e.Start(":" + config.App().Port))
 	}()
 
 	// graceful shutdown