+++ run.sh
@@ -3,7 +3,7 @@
 PROJECT="gochat"
 
 # run docker compose
-docker-compose up -d consul
+docker-compose up -d consul prometheus grafana
 
 # wait for consul container be ready
 while ! curl --request GET -sL --url 'http://localhost:8500/' > /dev/null 2>&1; do printf .; sleep 1; done